requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.6.0
async-timeout==4.0.3
attrs==22.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
distlib==0.3.6
Django==4.1.7
filelock==3.9.0
frozenlist==1.4.0
idna==3.4
iniconfig==2.0.0
magic-filter==1.0.12
multidict==6.0.4
packaging==23.0
pipenv==2023.5.19
platformdirs==3.0.0
pluggy==1.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pytest==7.2.2
python-dotenv==1.0.0
sqlparse==0.4.3
typing_extensions==4.7.1
virtualenv==20.19.0
virtualenv-clone==0.5.7
yarl==1.9.2
